requirements.txt
# Core dependencies
groq>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# Flight & Hotel APIs
amadeus>=9.0.0

# CLI enhancements
rich>=13.0.0

# Development
pytest>=8.0.0
